# Use ManuallyDrop instead of Option in Hole implementation