The Option is always Some until drop, where it becomes None. Use ManuallyDrop to avoid performing unnecessary unwraps, for a modest performance improvement. The use of ManuallyDrop requires Rust 1.20.0.

```
Before:
test pop_max_seq ... bench:      50,543 ns/iter (+/- 2,793)
test pop_min_seq ... bench:      74,096 ns/iter (+/- 3,004)
test push_seq    ... bench:      10,244 ns/iter (+/- 227)

After:
test pop_max_seq ... bench:      49,073 ns/iter (+/- 4,070)
test pop_min_seq ... bench:      71,585 ns/iter (+/- 3,480)
test push_seq    ... bench:       9,348 ns/iter (+/- 491)
```