@tov 2018-05-04
Nice. I didn't know about `ManuallyDrop`. Thanks!